requirements.txt
azure-cognitiveservices-vision-computervision==0.7.0
azure-common==1.1.28
azure-core==1.26.2
certifi==2022.12.7
charset-normalizer==3.0.1
contourpy==1.0.7
cycler==0.11.0
fonttools==4.38.0
idna==3.4
isodate==0.6.1
kiwisolver==1.4.4
matplotlib==3.6.3
msrest==0.7.1
numpy==1.24.1
oauthlib==3.2.2
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
typing-extensions==4.4.0
urllib3==1.26.14
